pyproject.toml
[project]
name = "serix"
version = "0.2.9"
description = "Automated adversarial testing for AI agents"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Yukta Thapliyal", email = "yukta@serixlabs.com"}]

keywords = [
    "security",
    "testing",
    "red-teaming",
    "fuzzing",
    "agents",
]

classifiers = [
    # MATURITY & AUDIENCE
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Typing :: Typed",
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "litellm>=1.0.0",
    "rich-gradient>=0.3.9",
    "tenacity>=8.2.0",
    "anthropic>=0.75.0",
    "google-generativeai>=0.8.6",
]

[project.urls]
Homepage = "https://github.com/yuktathapliyal/serix"
Repository = "https://github.com/yuktathapliyal/serix"
Documentation = "https://github.com/yuktathapliyal/serix#readme"

[project.scripts]
serix = "serix_v2.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "docs/",
    ".github/",
    "*.md",
    "!README.md",
    "serix.toml",
    "serix_vulnerable.toml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/serix", "src/serix_v2"]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests with mocked LLM responses",
    "real_api: tests that make real OpenAI API calls (requires OPENAI_API_KEY)",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=4.5.0",
    "respx>=0.21.0",
]
